requirements.txt
openai
pydub
python-telegram-bot
